runtime: new initRAMfromROM function, fix FPU init on noos/thumb
Separate the code that copies the data segment from ROM to RAM and
zeroes the remaining RAM area from rt0_go to its own function to allow
targets to use it or not (e.g. sometimes it's the bootloader job).

Always initialize FPU, even in softfloat mode. It allows to use a 32-bit
only FPU at least in assembly.