Rework behaviour of netpollBreak
It's not clear to me what should happen if netpollBreak is called with
no thread waiting in netpoll. The netpoll_stub.go implementation seems
to discard it, while other implementations seem to break the next
blocking call to netpoll.  Maybe it doesn't matter, maybe it's not
allowed.

Thus make netpollBreak callable from ISR, improving readability
and clear the netpollNote between sleeping and consuming the wakerq, to
ensure no wakeup is lost.