mips64: Fix disabling of interrupts
Disabling of interrupts was broken since after the call to sysirqctl(),
the STATUS register would be restored from exception context saved on
the stack. The globalIRQMask is now stored separately and always used to
restore the mask for external interrupts.

As a side-effect, the system will now boot with all external interrupts
disabled and waits for the user application to enable them.