mips64: Don't store REGTMP with GPRS
REGTMP is the scratch register used by the assembler. Saving it with
GPRS is bad because one needs to be extra careful writing assembly that
won't clobber REGTMP as long as it's not saved or after it's restored.

Instead try to save it as early and restore it as late as possible to
prevent bugs, that will only be visible in the disassembly.