[merge go1.24] disable support for FIPS 140 on noos and thumb
If enabled it changes the relocation type of  moduledata.maxpc from STEXT to
STEXTFIPSEND which in effect removes the thumb bit from the maxpc. There's
probably a better way to handle this problem but for now we don't need FIPS
on noos or thumb so we chose the simplest way.